+++ CHANGELOG.md
@@ -6,13 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- **deps**: typescript-eslint (dev dep)
+
 ### Improved
 - **deps**: Upgrade eslint to 9.11.1
 - **deps**: Upgrade nyc to 17.1.0
 - **deps**: Upgrade typescript to 5.6.2
 - **deps**: Upgrade @types/node to 20.16.10
-- **deps**: Upgrade @typescript-eslint/eslint-plugin to 8.7.0
-- **deps**: Upgrade @typescript-eslint/parser to 8.7.0
+
+### Removed
+- **deps**: @typescript-eslint/eslint-plugin
+- **deps**: @typescript-eslint/parser
 
 ## [12.23.0] - 2024-09-16
 ### Added