+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Improved
+- **deps**: Upgrade eslint to 9.12.0
+- **deps**: Upgrade typescript-eslint to 8.8.0
+
 ## [12.24.0] - 2024-09-29
 ### Added
 - **deps**: typescript-eslint (dev dep)